为EEPROM读写增加数据切片功能，保证读写大量数据时的准确性 修改dataSize参数类型为uint8_t

+++ Code/Drivers/Hardware/Inc/driver_at24c0x.h
@@ -3,11 +3,13 @@
  *  @details    支持从AT24C0X指定页读写的不限类型不限长度的数据
  *  @author     Harry-Qu
  *  @date       2022/11/5
- *  @version    1.1
+ *  @version    1.1.1
  *  @par        日志
  *              1.0     |   实现AT24C0x基本读写功能
  *              1.1     |   新增从指定页中读写数据的接口driver_at24c0x_writeData_page, driver_at24c0x_readData_page
  *                          修改部分接口返回类型为uint8_t
+ *              1.1.1   |   为ROM读写增加数据切片功能，保证读写大量数据时的准确性
+ *                          修改dataSize参数类型为uint8_t
 */
 
 #ifndef DRIVER_AT24C0X_H
@@ -30,7 +32,7 @@
  *         0: success
  *         1: fail
  */
-uint8_t driver_at24c0x_writeData(uint8_t address, void *data, uint32_t dataSize);
+uint8_t driver_at24c0x_writeData(uint8_t address, void *data, uint8_t dataSize);
 
 /**
  * 从AT24C0X中读取数据
@@ -41,7 +43,7 @@ uint8_t driver_at24c0x_writeData(uint8_t address, void *data, uint32_t dataSize)
  *         0: success
  *         1: fail
  */
-uint8_t driver_at24c0x_readData(uint8_t address, void *data, uint32_t dataSize);
+uint8_t driver_at24c0x_readData(uint8_t address, void *data, uint8_t dataSize);
 
 /**
  * 向AT24C0X指定页中写入数据
@@ -53,7 +55,7 @@ uint8_t driver_at24c0x_readData(uint8_t address, void *data, uint32_t dataSize);
  *         0: success
  *         1: fail
  */
-uint8_t driver_at24c0x_writeData_page(uint8_t page, uint8_t address, void *data, uint32_t dataSize);
+uint8_t driver_at24c0x_writeData_page(uint8_t page, uint8_t address, void *data, uint8_t dataSize);
 
 /**
  * 从AT24C0X指定页中读取数据
@@ -65,6 +67,6 @@ uint8_t driver_at24c0x_writeData_page(uint8_t page, uint8_t address, void *data,
  *         0: success
  *         1: fail
  */
-uint8_t driver_at24c0x_readData_page(uint8_t page, uint8_t address, void *data, uint32_t dataSize);
+uint8_t driver_at24c0x_readData_page(uint8_t page, uint8_t address, void *data, uint8_t dataSize);
 
 #endif //DRIVER_AT24C0X_H
